move gyro handlers to its own file
