+++ src/mappings/poolFactory.ts
@@ -35,18 +35,13 @@ import { ConvergentCurvePool as CCPoolTemplate } from '../types/templates';
 import { LiquidityBootstrappingPool as LiquidityBootstrappingPoolTemplate } from '../types/templates';
 import { InvestmentPool as InvestmentPoolTemplate } from '../types/templates';
 import { LinearPool as LinearPoolTemplate } from '../types/templates';
-import { Gyro2Pool as Gyro2PoolTemplate } from '../types/templates';
-import { Gyro3Pool as Gyro3PoolTemplate } from '../types/templates';
-import { GyroEPool as GyroEPoolTemplate } from '../types/templates';
+
 import { FXPool as FXPoolTemplate } from '../types/templates';
 
 import { WeightedPool } from '../types/templates/WeightedPool/WeightedPool';
 import { StablePool } from '../types/templates/StablePool/StablePool';
 import { ConvergentCurvePool } from '../types/templates/ConvergentCurvePool/ConvergentCurvePool';
 import { LinearPool } from '../types/templates/LinearPool/LinearPool';
-import { Gyro2Pool } from '../types/templates/Gyro2Pool/Gyro2Pool';
-import { Gyro3Pool } from '../types/templates/Gyro3Pool/Gyro3Pool';
-import { GyroEPool } from '../types/templates/GyroEPool/GyroEPool';
 import { ERC20 } from '../types/Vault/ERC20';
 
 function createWeightedLikePool(event: PoolCreated, poolType: string, poolTypeVersion: i32 = 1): string | null {
@@ -356,122 +351,6 @@ function handleNewLinearPool(event: PoolCreated, poolType: string, poolTypeVersi
   LinearPoolTemplate.create(poolAddress);
 }
 
-export function handleNewGyro2Pool(event: PoolCreated): void {
-  let poolAddress: Address = event.params.pool;
-
-  let poolContract = Gyro2Pool.bind(poolAddress);
-
-  let poolIdCall = poolContract.try_getPoolId();
-  let poolId = poolIdCall.value;
-
-  let swapFeeCall = poolContract.try_getSwapFeePercentage();
-  let swapFee = swapFeeCall.value;
-
-  let pool = handleNewPool(event, poolId, swapFee);
-
-  pool.poolType = PoolType.Gyro2;
-  let sqrtParamsCall = poolContract.try_getSqrtParameters();
-  pool.sqrtAlpha = scaleDown(sqrtParamsCall.value[0], 18);
-  pool.sqrtBeta = scaleDown(sqrtParamsCall.value[1], 18);
-
-  let tokens = getPoolTokens(poolId);
-  if (tokens == null) return;
-  pool.tokensList = tokens;
-
-  pool.save();
-
-  handleNewPoolTokens(pool, tokens);
-
-  Gyro2PoolTemplate.create(event.params.pool);
-}
-
-export function handleNewGyro3Pool(event: PoolCreated): void {
-  let poolAddress: Address = event.params.pool;
-
-  let poolContract = Gyro3Pool.bind(poolAddress);
-
-  let poolIdCall = poolContract.try_getPoolId();
-  let poolId = poolIdCall.value;
-
-  let swapFeeCall = poolContract.try_getSwapFeePercentage();
-  let swapFee = swapFeeCall.value;
-
-  let pool = handleNewPool(event, poolId, swapFee);
-
-  pool.poolType = PoolType.Gyro3;
-  let root3AlphaCall = poolContract.try_getRoot3Alpha();
-
-  if (!root3AlphaCall.reverted) {
-    pool.root3Alpha = scaleDown(root3AlphaCall.value, 18);
-  }
-
-  let tokens = getPoolTokens(poolId);
-  if (tokens == null) return;
-  pool.tokensList = tokens;
-
-  pool.save();
-
-  handleNewPoolTokens(pool, tokens);
-
-  Gyro3PoolTemplate.create(event.params.pool);
-}
-
-function createGyroEPool(event: PoolCreated, poolTypeVersion: i32 = 1): void {
-  let poolAddress: Address = event.params.pool;
-  let poolContract = GyroEPool.bind(poolAddress);
-
-  let poolIdCall = poolContract.try_getPoolId();
-  let poolId = poolIdCall.value;
-
-  let swapFeeCall = poolContract.try_getSwapFeePercentage();
-  let swapFee = swapFeeCall.value;
-
-  let pool = handleNewPool(event, poolId, swapFee);
-
-  pool.poolType = PoolType.GyroE;
-  pool.poolTypeVersion = poolTypeVersion;
-  let eParamsCall = poolContract.try_getECLPParams();
-
-  if (!eParamsCall.reverted) {
-    const params = eParamsCall.value.value0;
-    const derived = eParamsCall.value.value1;
-    pool.alpha = scaleDown(params.alpha, 18);
-    pool.beta = scaleDown(params.beta, 18);
-    pool.c = scaleDown(params.c, 18);
-    pool.s = scaleDown(params.s, 18);
-    pool.lambda = scaleDown(params.lambda, 18);
-
-    // terms in the 'derived' object are stored in extra precision (38 decimals) with final decimal rounded down
-    pool.tauAlphaX = scaleDown(derived.tauAlpha.x, 38);
-    pool.tauAlphaY = scaleDown(derived.tauAlpha.y, 38);
-    pool.tauBetaX = scaleDown(derived.tauBeta.x, 38);
-    pool.tauBetaY = scaleDown(derived.tauBeta.y, 38);
-    pool.u = scaleDown(derived.u, 38);
-    pool.v = scaleDown(derived.v, 38);
-    pool.w = scaleDown(derived.w, 38);
-    pool.z = scaleDown(derived.z, 38);
-    pool.dSq = scaleDown(derived.dSq, 38);
-  }
-
-  let tokens = getPoolTokens(poolId);
-  if (tokens == null) return;
-  pool.tokensList = tokens;
-
-  pool.save();
-
-  handleNewPoolTokens(pool, tokens);
-
-  GyroEPoolTemplate.create(event.params.pool);
-}
-
-export function handleNewGyroEPool(event: PoolCreated): void {
-  createGyroEPool(event);
-}
-
-export function handleNewGyroEV2Pool(event: PoolCreated): void {
-  createGyroEPool(event, 2);
-}
-
 export function handleNewFXPool(event: ethereum.Event): void {
   /**
    * FXPoolFactory emits a custom NewFXPool event with the following params:
@@ -529,7 +408,7 @@ function findOrInitializeVault(): Balancer {
   return vault;
 }
 
-function handleNewPool(event: PoolCreated, poolId: Bytes, swapFee: BigInt): Pool {
+export function handleNewPool(event: PoolCreated, poolId: Bytes, swapFee: BigInt): Pool {
   let poolAddress: Address = event.params.pool;
 
   let pool = Pool.load(poolId.toHexString());
@@ -577,7 +456,7 @@ function handleNewPool(event: PoolCreated, poolId: Bytes, swapFee: BigInt): Pool
   return pool;
 }
 
-function handleNewPoolTokens(pool: Pool, tokens: Bytes[]): void {
+export function handleNewPoolTokens(pool: Pool, tokens: Bytes[]): void {
   let tokensAddresses = changetype<Address[]>(tokens);
 
   for (let i: i32 = 0; i < tokens.length; i++) {
